development/services/maps-access-service/requirements.txt
# Maps Access Service Dependencies

# FastAPI and ASGI server
fastapi>=0.115.0
starlette>=0.52.0
uvicorn[standard]>=0.27.0

# HTTP client (Google Maps API calls)
httpx>=0.27.0

# Pydantic settings
pydantic-settings>=2.0.0

# Redis (geocode caching)
redis>=5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0  # Parallel test execution
# Structured logging and metrics
python-json-logger>=2.0.0

# Prometheus metrics
prometheus-client>=0.19.0